working mac and windows

+++ epic_audio.py
@@ -1,24 +1,31 @@
-import pyaudio,wave,sys
+import wave,sys
+
+try:
+    import pyaudio
+    USE_PYAUDIO = True
+except ImportError:
+    USE_PYAUDIO = False
 
 def playSound(filename):
-    CHUNK = 1024
+    if USE_PYAUDIO:
+        CHUNK = 1024
 
-    wf = wave.open(filename, 'rb')
+        wf = wave.open(filename, 'rb')
 
-    p = pyaudio.PyAudio()
+        p = pyaudio.PyAudio()
 
-    stream = p.open(format=p.get_format_from_width(wf.getsampwidth()),
-                    channels=wf.getnchannels(),
-                    rate=wf.getframerate(),
-                    output=True)
+        stream = p.open(format=p.get_format_from_width(wf.getsampwidth()),
+                        channels=wf.getnchannels(),
+                        rate=wf.getframerate(),
+                        output=True)
 
-    data = wf.readframes(CHUNK)
-
-    while data != '':
-        stream.write(data)
         data = wf.readframes(CHUNK)
 
-    stream.stop_stream()
-    stream.close()
+        while data != '':
+            stream.write(data)
+            data = wf.readframes(CHUNK)
 
-    p.terminate()
+        stream.stop_stream()
+        stream.close()
+
+        p.terminate()
